feat: create new authModule
