+++ app/backend/src/modules/api.module.js
@@ -1,9 +1,9 @@
 import fastifyRateLimit from "@fastify/rate-limit";
+import authModule from "./auth.module.js";
 import matchRoutes from "../routes/matches.routes.js";
 import tournamentRoutes from "../routes/tournaments.routes.js";
 import userstatsRoutes from "../routes/user_stats.routes.js";
 import userRoutes from "../routes/users.routes.js";
-import authRoutes from "../routes/auth.routes.js";
 import env from "../config/env.js";
 
 export default async function apiModule(fastify) {
@@ -27,9 +27,10 @@ export default async function apiModule(fastify) {
     }
   });
 
+  await fastify.register(authModule);
+
   await fastify.register(userRoutes, { prefix: "/users" });
   await fastify.register(matchRoutes, { prefix: "/matches" });
   await fastify.register(tournamentRoutes, { prefix: "/tournaments" });
   await fastify.register(userstatsRoutes, { prefix: "/user-stats" });
-  await fastify.register(authRoutes, { prefix: "/auth" });
 }
