+++ app/backend/src/routes/auth.routes.js
@@ -8,28 +8,8 @@ import {
 import { errorResponses } from "../utils/error.js";
 import { authorizeUserAccess } from "../middleware/auth.js";
 import env from "../config/env.js";
-import fastifyRateLimit from "@fastify/rate-limit";
 
 export default async function authRoutes(fastify) {
-  await fastify.register(fastifyRateLimit, {
-    max: env.authRateLimitMax,
-    timeWindow: env.authRateLimitTimeInMS,
-    hook: "preHandler",
-    keyGenerator: function (request) {
-      return request.user?.id ?? request.ip;
-    },
-    onExceeded: (request) => {
-      if (request.user?.id) {
-        request.log.warn(
-          `Rate limit exceeded for user id: ${request.user.id} - URL: ${request.raw.url}`
-        );
-      } else {
-        request.log.warn(
-          `Rate limit exceeded for IP: ${request.ip} - URL: ${request.raw.url}`
-        );
-      }
-    }
-  });
   fastify.post("/login", optionsloginUser, loginUserHandler);
 
   fastify.get("/token", optionsAuthUserAccess, authAndDecodeAccessHandler);
